% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_helpers.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\alias{scrape_func}
\alias{season_only_sources}
\alias{playoff_sources}
\title{Helper objects for data sources}
\description{
\describe{
  \item{\code{scrape_func}}{A list with the names of the functions used to
  scrape data from a source}
  \item{\code{season_only_sources}}{Vector with names of sources that only
  provide seasonal data}
  \item{\code{playoff_sources}}{Vector with names of sources that
  provide data for playoffs (week 17-20)}
}
}
\keyword{datasets}
