% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ff_player_data}
\alias{ff_player_data}
\title{Player data from MyFantasyLeague}
\format{A data frame with 2166 rows and 6 columns.
\describe{
  \item{id}{Player's ID at MyFantasyLeague}
  \item{name}{Player's name}
  \item{position}{Player's position as listed at MyFantasyLeague}
  \item{team}{Player's team as listed at MyFantasyLeague}
  \item{draft_year}{Year player was drafted}
  \item{birthdate}{Player's birthdate}
}}
\usage{
ff_player_data
}
\description{
A dataset containing player data from MyFantasyLeague. This dataset is generated
from the MyFantasyLeague API.
}
\keyword{datasets}
