% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_salaries.R
\name{get_salaries}
\alias{get_salaries}
\title{Retrieve salaries for a specific site}
\usage{
get_salaries(site = c("fanduel", "draftkings", "yahoo", "fantasydraft",
  "fantasyscore"))
}
\arguments{
\item{site}{Specifies the site to retrieve salaries from. Should be one of
\code{c("fanduel", "draftkings", "yahoo", "fantasydraft", "fantasyscore")}.
If ommited then FanDuel salaries will be scraped.}
}
\value{
A \link{data.frame} with salary data.
}
\description{
This function retrieves the player salaries for a site.
}
