% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_funcs.R
\name{Impute}
\alias{Impute}
\alias{pass_impute}
\alias{rush_impute}
\alias{rec_impute}
\alias{kick_impute}
\title{Imputing values for missing data}
\usage{
pass_impute(pass_tbl)

rush_impute(rush_tbl)

rec_impute(rec_tbl)

kick_impute(kick_tbl)
}
\description{
Differnt sources provide different data. The functions below will fill out
NA values for some sources with values based on data from other sources
}
