% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cbs.R
\name{scrape_cbs}
\alias{scrape_cbs}
\title{Scrape data from CBS}
\usage{
scrape_cbs(week = NULL, position = c("QB", "RB", "WR", "TE", "K", "DST"))
}
\arguments{
\item{week}{Week number that data will be scraped for. If \code{= 0} or omitted
season data will be scraped}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST")}. If omitted QB data will be scraped.}
}
\value{
A \link{data.table} with results from the scrape.
}
\description{
Use this function to scrape fantasy football projections from CBS.
}
