% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_data.R
\name{scrape_data}
\alias{scrape_data}
\title{Scrape data from multiple sources}
\usage{
scrape_data(season = NULL, week = 0, position = c("QB", "RB", "WR", "TE",
  "K", "DST", "DL", "LB", "DB"), sources = c("CBS", "Yahoo", "ESPN",
  "FantasyData", "FantasyPros", "FantasySharks", "FFToday", "FleaFlicker",
  "NFL", "NumberFire", "RTSports", "Walterfootball"))
}
\description{
Scrape data from multiple sources
}

