% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ecr.R
\name{scrape_ecr}
\alias{scrape_ecr}
\title{Scrape FantasyPro rankings}
\usage{
scrape_ecr(rank_period = c("draft", "weekly", "ros", "dynasty"),
  position = c("Overall", "QB", "RB", "WR", "TE", "K", "FLEX", "DST", "IDP",
  "DL", "LB", "DB"), rank_type = c("Std", "PPR", "Half"))
}
\arguments{
\item{rank_period}{Specify the period rankings are being scraped for. Should
be one of \code{c("draft", "weekly", "ros", "dynasty")}. If omitted then draft
rankings will be scraped}

\item{position}{The position that rankings should be scraped for. Should be one
of \code{c("Overall", "QB", "RB", "WR", "TE", "K", "FLEX", "DST", "IDP", "DL", "LB", "DB")}.
If not specified then Overall rankings will be scraped.}

\item{rank_type}{The type of rankings to be scraped. Should be one of
\code{c("Std", "PPR", "Half")}. If not speficied then Standard rankings will be
scraped.}
}
\description{
Use this function to scrape Expert Consensus Rankings (ECR) from FantasyPros
}
