% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_espn.R
\name{scrape_espn}
\alias{scrape_espn}
\title{Scrape data from ESPN}
\usage{
scrape_espn(season, week = NULL, position = c("QB", "RB", "WR", "TE", "DST",
  "K"))
}
\arguments{
\item{season}{The year that data will be scraped for}

\item{week}{The week that data will be scraped for. If \code{= 0} or omitted
season data will be scraped}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST")}. If omitted QB data will be scraped.}
}
\description{
Use this function to srape fantasy football projections from ESPN
}
