% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nerd.R
\name{scrape_fantasynerd}
\alias{scrape_fantasynerd}
\title{Scrape data from FantasyFootballNerd}
\usage{
scrape_fantasynerd(week = NULL, position = c("QB", "RB", "WR", "TE", "K",
  "DST"))
}
\arguments{
\item{week}{The week that data will be scraped for. If \code{= 0} or omitted
season data will be scraped}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST")}. If omitted QB data will be scraped.}
}
\description{
Use this function to srape fantasy football projections from FantasyFootballNerd.
The function uses the FantasyFootball Nerd API. If you don't have an API key
to use for this you will not get accuracte results. You can register for an
API key at fantasyfootballnerd.com. 
You set the API key for this scrape via \code{options('ffdata.ffn_api' = 'apikey')}
}
