% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fantasyshark.R
\name{scrape_fantasysharks}
\alias{scrape_fantasysharks}
\title{Scrape data from FantasySharks}
\usage{
scrape_fantasysharks(season = NULL, week = NULL, position = c("QB", "RB",
  "WR", "TE", "K", "DEF", "DL", "LB", "DB"))
}
\arguments{
\item{season}{The year that data will be scraped for. If ommitted the current 
season will be used}

\item{week}{The week that data will be scraped for. If \code{= 0} or omitted
season data will be scraped}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST")}. If omitted QB data will be scraped.}
}
\description{
Use this function to srape fantasy football projections from FantasySharks
}
