% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_fftoday.R
\name{scrape_fftoday}
\alias{scrape_fftoday}
\title{Scrape data from FFToday}
\usage{
scrape_fftoday(season = NULL, week = NULL, position = c("QB", "RB", "WR",
  "TE", "K", "DEF", "DL", "LB", "DB"))
}
\description{
Scrape data from FFToday
}

