% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_flea.R
\name{scrape_fleaflick}
\alias{scrape_fleaflick}
\title{Scrape data from FleaFlicker}
\usage{
scrape_fleaflick(week = NULL, position = c("QB", "RB", "WR", "TE", "Flex",
  "K", "DST", "DB", "DL", "LB", "IDP"))
}
\arguments{
\item{week}{The week that data will be scraped for. If omitted, season data
will be scraped.}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "Flex", "K", "DST", "DB", "DL", "LB", "IDP")}.
If omitted QB data will be scraped. Specifying \code{"Flex"} will scrape RB,
WR, and TE data. Specifying \code{"IDP"} will scrape DB, DL, and LB data.}
}
\description{
Use this function to srape fantasy football projections from FleaFlicker
}
