% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nfl.R
\name{scrape_nfl}
\alias{scrape_nfl}
\title{Scrape data from NFL}
\usage{
scrape_nfl(season = NULL, week = NULL, position = c("QB", "RB", "WR",
  "TE", "K", "DST", "DL", "LB", "DB"))
}
\arguments{
\item{season}{The year that data will be scraped for. If omitted the current
season will scraped.}

\item{week}{The week that data will be scraped for. If \code{= 0} or omitted
season data will be scraped.}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST" "DL", "LB", "DB")}. If omitted QB data will be scraped.}
}
\description{
Use this function to srape fantasy football projections from NFL. This function
is using the NFL Fantasy api.
}

