% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_numfire.R
\name{scrape_numfire}
\alias{scrape_numfire}
\title{Scrape data from NumberFire}
\usage{
scrape_numfire(week = NULL, position = c("Off", "QB", "RB", "WR", "TE",
  "RB/WR", "K", "DST", "IDP"))
}
\arguments{
\item{week}{The week that data will be scraped for. If \code{= 0} or omitted
season data will be scraped}

\item{position}{The player position to scrape data for. Has to be one of
\code{c("Off", "QB", "RB", "WR", "TE", "RB/WR", "K", "DST", "IDP")}. If omitted 
data for all offensive positions (QB, RB, WR, and TE) will be scraped.}
}
\description{
Use this function to srape fantasy football projections from NumberFire
}
