% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rts.R
\name{scrape_rts}
\alias{scrape_rts}
\title{Scrape data from RTSports}
\usage{
scrape_rts(position = c("QB", "RB", "WR", "TE", "K", "DST"))
}
\arguments{
\item{position}{The player position to scrape data for. Has to be one of
\code{c("QB", "RB", "WR", "TE", "K", "DST")}. If omitted QB data will be scraped.}
}
\description{
Use this function to srape fantasy football projections from RTSports. Only
seasonal projections are available.
}
