% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_wf.R
\name{scrape_wf}
\alias{scrape_wf}
\title{Scrape data from WalterFootball}
\usage{
scrape_wf(season = NULL)
}
\arguments{
\item{season}{The season to scrape data for.}
}
\description{
Use this function to srape fantasy football projections from WalterFootball Only
seasonal projections are available.
}
